split table button (required major change!)
Mostly because tables (experiments) in brainspell were indexed by their
position in the article. The split button required to index experiments
by a unique experiment id – eid – which allows to keep tags assigned to
the correct experiment despite their position in the table.
I added a new part to the locations tables, visible only to logged-in
users, that hosts the Split button, and will soon host some other
buttons.
There are 3 php scripts that change the organisation of the database
from iExp to EID, they are in the php directory (kept mostly for
historical reasons, because they shouldn’t be used again)
* I also eliminated all CDATA tags from the main database, and I’m now
using just a mix of json_encode/decode and mysql string encoding.